pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/distributed_downloader"]

[project]
name = "distributed-downloader"
dynamic = ["version"]
authors = [
    { name = "Andrey Kopanev", email = "kopanev.1@osu.edu" },
    { name = "Elizabeth G. Campolongo", email = "e.campolongo479@gmail.com" },
    { name = "Matthew J. Thompson", email = "thompson.m.j@outlook.com" },
]
description = "A tool for downloading files from a list of URLs in parallel."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Brotli",
    "certifi",
    "charset-normalizer",
    "Cython",
    "h5py",
    "hdf5plugin",
    "idna",
    "inflate64",
    "mpi4py",
    "multivolumefile",
    "numpy",
    "opencv-python",
    "pandas",
    "pillow",
    "psutil",
    "py4j",
    "py7zr",
    "pybcj",
    "pycryptodomex",
    "pyppmd",
    "pyspark",
    "python-dateutil",
    "python-dotenv",
    "pytz",
    "pyzstd",
    "requests",
    "six",
    "texttable",
    "tzdata",
    "urllib3",
]

[project.optional-dependencies]
dev = ["pytest"]

keywords = [
    "parallel",
    "distributed",
    "download",
    "url",
]

[project.urls]
Homepage = "https://github.com/Imageomics/distributed-downloader"
Repository = "https://github.com/Imageomics/distributed-downloader.git"
"Bug Tracker" = "https://github.com/Imageomics/distributed-downloader/issues"

[project.scripts]
distributed_downloader = "distributed_downloader.main:main"

[tool.hatch.version]
path = "src/distributed_downloader/__about__.py"